requirements.txt
streamlit>=1.0
spacy>=3.0
pymupdf>=1.19.0
python-docx>=0.8.11
pandas>=1.3
regex
tqdm